requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-environ==0.12.0
django-jazzmin==3.0.1
django-leaflet==0.33.0
geographiclib==2.1
geopy==2.4.1
idna==3.11
numpy==2.4.2
pandas==3.0.0
polars-runtime-32==1.38.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
